requirements.txt
Flask==2.2.2
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
Pillow==9.3.0
requests==2.28.1
regex==2022.10.31